update after second self-review
